Made Makefile more complicated
There are now additional targets such as install-nolinks, install-parts
and so on. The main reason is that one can now install the program
without having to install the symlinks in /etc/network and the side
effect is that make can now be run with -j (which is pointless for such
a small installation, but nevermind).